Adapt to upstream change in reviewer suggestion result
Previously the result would contain a list of maps, where the maps
would have entries with the key being either "account" or "group"
and the value being AccountInfo or GroupBaseInfo. Both of the Info
classes have a 'name' member.

Since change Id1323c5a0 the maps may also include an integer 'count'
and boolean 'confirm'. Calling get('name') on those returns null, and
then the subsequent call to 'toString()' fails.

Move the calls to get('name') into the conditional blocks where we
know the item is AccountInfo or GroupBaseInfo.

Bug: Issue 7008
Change-Id: Ic00927eb2e9c86efc71452289c89aaff702ba07e