Handle when uploader email matches more than one account
It's possible for an email to be used by more than one account. When the
uploader email matches more than one account, ChangeEventListener was
throwing an IllegalArgumentException preventing the operation from
completing successfully.

In most of the case, this is not harmful since ChangeEvent listeners
are called at the end of the operation but in the following scenario, it
makes gerrit data inconsistent.

If user push a new patchset to an existing change but bypass review, the
operation will be aborted by the IllegalArgumentException resulting in
the change being merged in the repo but its status still open.

Change-Id: Icc48fe4548fc46169bec3561ca490a199107af3b